Cleaning up UI and Bindings access
Basically the goal here is to make it easier to access things in a clean way. The Denhac UI struct helps keep instances of all of the window classes so that any window can access any other in a static way. DenhacBindings provides a cleaner way (at least from the context of the DenhacBoard class) to bind all of the callbacks to where they need to go.