Improving SiteLink protocol slightly
This diff does two things. First we reduce the size of a product update
by shortening the max length of the name. This ensures that the max product
packet length is 63 bytes.

The second thing it does is not send product updates unless the host requests
it. This is to work around some of the issues with ACKing packets.