Adjusting how ACKs work for fetching orders
This is to help prevent buffer overflowing and will be one of several
diffs like it. The basic concept is, if the next piece of the packet,
whether it's the order header, item, etc. would overflow the buffer,
then we wait for an ACK from the other side. Both sides use the same
math so the ACK and wait for ACK on the same schedule.