Added Order Cancelled function to SiteLink
I had to fix the WiFi bridge too because the cancel endpoint doesn't
return a body so the JSON decoding fails. Now the bridge grabs the
content length from the response in the event we ever need that again.

If we need to grab more of these things from the API endpoint, we should
look into regex somehow.

Next up is the last existing endpoint which is the "update order" one.
Primarily this will be used to verify someone really can vend what they
want AND to update vended amounts after the vend itself has kicked off.