Inverted the colors on the main window buttons
I originally made the buttons have the color border and the white
background when not pressed and then switched to one solid color
with white text when pressed. This diff inverts that.

If someone is colorblind then the enabled not-pressed button and
the disabled button would look basically the same. Plus this
honestly just looks better with the pressed action. In real life
you would think of a button being pushed down and this gives a
similar effect without the use of shadows.

I did modify the row/col buttons here but they don't match that
pattern. I did that for two reasons. One is that when it's a solid
background, it tends to be overpowering. Part of that I think is
that I used 1px black lines to separate the buttons so it all just
blends together and doesn't look good. The other aspect is that you
have a visual indication of the item you want to vend because you'll
see it in the machine. Eventually we'll need to check product stock
levels too and we might end up with the case that an item is visually
there in the machine but the button should be disabled. That may
prove confusing for people, but they'll probably just assume we
miscounted the items in the machine.