Full vend functionality works
It's very basic and doesn't handle corner cases, but it was very
cool to watch the order update to the 'processing' status and then
update to the 'vending' status.

I still haven't done this with it all installed in the machine, so
no idea if it actually turns the motors with this version. We've
made some heavy modifications to that class since the last test.